Applies new clippy suggestions

+++ y2020/ex13/src/lib.rs
@@ -59,7 +59,7 @@ pub fn part2(input: &str) -> u64 {
         candidate += step;
         // now we check if this creates a scale of timestamps
         for (index, bus) in buses.iter().enumerate() {
-            if (candidate + bus.offset as u64) % bus.id as u64 == 0 {
+            if (candidate + bus.offset) % bus.id == 0 {
                 step *= bus.id; // here we should have done a LCM, but the bus ids are all prime, so...
                 satisfied_buses.push(index);
             }
